string_util.c/h: - is_abbrev(), str_starts_with():   - return false for degenerate cases where input parms are null - is_abbrev():   - change type of minchars from int to size_t to a) avoid negative values and b) conform to memcmp() signature - documentation and cosmetic

+++ src/util/string_util.h
@@ -30,6 +30,7 @@
 
 /** \cond */
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 #include <stdio.h>
 /** \endcond */
@@ -60,7 +61,7 @@ char * bool_repr(int value);
 //
 
 bool   streq(const char * s1, const char * s2);
-bool   is_abbrev(const char * value, const char * longname, int minchars);
+bool   is_abbrev(const char * value, const char * longname, size_t minchars);
 bool   str_starts_with(const char * value_to_test, const char * start_part);
 bool   str_ends_with(const char * value_to_test, const char * end_part);
 bool   str_all_printable(const char * s);
